perf: don't load codemirror on home page
