+++ specifyweb/frontend/js_src/lib/components/AppResources/codeMirrorLinters.ts
@@ -15,6 +15,7 @@ import { findNodePosition } from '../Syncer/findNodePosition';
 import { syncers } from '../Syncer/syncers';
 import type { SimpleXmlNode } from '../Syncer/xmlToJson';
 import { toSimpleXmlNode, xmlToJson } from '../Syncer/xmlToJson';
+import { parseXml } from './parseXml';
 
 export const createLinter =
   (handler: (view: EditorView) => RA<Diagnostic>) =>
@@ -72,33 +73,6 @@ function parseXmlUsingSpec(
 }
 
 export const jsonLinter = createLinter(jsonParseLinter());
-
-export function parseXml(string: string): Element | string {
-  const parsedXml = new globalThis.DOMParser().parseFromString(
-    string,
-    'text/xml'
-  ).documentElement;
-
-  // Chrome, Safari
-  const parseError = parsedXml.getElementsByTagName('parsererror')[0];
-  if (typeof parseError === 'object')
-    return (parseError.children[1].textContent ?? parseError.innerHTML).trim();
-  // Firefox
-  else if (parsedXml.tagName === 'parsererror')
-    return (
-      parsedXml.childNodes[0].nodeValue ??
-      parsedXml.textContent ??
-      parsedXml.innerHTML
-    ).trim();
-  else return parsedXml;
-}
-
-export function strictParseXml(xml: string): Element {
-  const parsed = parseXml(xml);
-  // eslint-disable-next-line functional/no-throw-statement
-  if (typeof parsed === 'string') throw new Error(parsed);
-  else return parsed;
-}
 
 const xmlErrorParsers = [
   /(?<message>[^\n]+)\n[^\n]+\nLine Number (?<line>\d+), Column (?<column>\d+)/u,
