+++ specifyweb/frontend/js_src/lib/components/FormEditor/__tests__/viewSpec.test.ts
@@ -1,6 +1,6 @@
 import { requireContext } from '../../../tests/helpers';
 import { localized } from '../../../utils/types';
-import { strictParseXml } from '../../AppResources/codeMirrorLinters';
+import { strictParseXml } from '../../AppResources/parseXml';
 import { tables } from '../../DataModel/tables';
 import { error } from '../../Errors/assert';
 import { formatXmlForTests } from '../../Syncer/__tests__/utils';
