+++ specifyweb/frontend/js_src/lib/components/FormParse/__tests__/commands.test.ts
@@ -1,5 +1,5 @@
 import { requireContext } from '../../../tests/helpers';
-import { strictParseXml } from '../../AppResources/codeMirrorLinters';
+import { strictParseXml } from '../../AppResources/parseXml';
 import { tables } from '../../DataModel/tables';
 import type { SimpleXmlNode } from '../../Syncer/xmlToJson';
 import { toSimpleXmlNode, xmlToJson } from '../../Syncer/xmlToJson';
