+++ specifyweb/frontend/js_src/lib/components/FormParse/__tests__/fields.test.ts
@@ -1,7 +1,7 @@
 import { reportsText } from '../../../localization/report';
 import { requireContext } from '../../../tests/helpers';
 import { localized } from '../../../utils/types';
-import { strictParseXml } from '../../AppResources/codeMirrorLinters';
+import { strictParseXml } from '../../AppResources/parseXml';
 import { getField } from '../../DataModel/helpers';
 import { tables } from '../../DataModel/tables';
 import type { SimpleXmlNode } from '../../Syncer/xmlToJson';
