+++ specifyweb/frontend/js_src/lib/components/FormParse/__tests__/index.test.ts
@@ -5,7 +5,7 @@ import { Http } from '../../../utils/ajax/definitions';
 import type { RA } from '../../../utils/types';
 import { ensure, localized } from '../../../utils/types';
 import { removeKey } from '../../../utils/utils';
-import { strictParseXml } from '../../AppResources/codeMirrorLinters';
+import { strictParseXml } from '../../AppResources/parseXml';
 import { tables } from '../../DataModel/tables';
 import { getPref } from '../../InitialContext/remotePrefs';
 import { formatUrl } from '../../Router/queryString';
